examples/turkish-financial-data-scraper/requirements.txt
# Firecrawl
firecrawl-py>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Data Processing
pandas>=2.1.4
numpy<2.0.0

# PDF Processing
pdfplumber>=0.10.3
PyPDF2>=3.0.1
PyMuPDF>=1.23.0  # fitz for PDF text extraction
fpdf2>=2.7.0  # PDF generation with Unicode support

# LLM Integration
openai>=1.3.0  # Works with OpenAI and local LM Studio

# HTTP & Web
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
aiohttp>=3.9.0  # Async HTTP client

# Scheduling
apscheduler>=3.10.4

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.2

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Logging
colorlog>=6.8.0

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0
